Avoid lib:sendw which blindly accepts messages.
Trying to test enif_send() semantics was causing some weird errors in
etap when it waited to see if a test should be skipped. The use of
lib:sendw/2 was eating a message without paying attention to what it
was. This refactors things a bit to include an {etap, Mesg} envelope
around replies from the test server.